Remove const from sequences to be modified
Do not assume the sequence won't need to be modified when passing it to
various functions, and drop the `const` qualifier from the corresponding
parameters. This allows us to use a `container_of` style macro to reach
for private data, instead of using the `data` member, and thus avoid
casting from const member to non-const container.